Turn status bar feature into its own class
